requirements.txt
pip==23.3.2
asgiref==3.7.2
crispy-bootstrap4==2023.1
dj-database-url==2.1.0
Django ==5.0
django-crispy-forms==2.1
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
djangorestframework~=3.14.0
opencv-python==4.9.0.80
numpy==1.26.2
channels==4.0.0
channels-redis~=4.1.0
django-cors-headers~=4.3.1
whitenoise==6.6.0
playwright~=1.40.0
# django_schema_viewer==0.4.1